utils/requirements.txt
boto3==1.5.1
botocore==1.8.15
certifi==2017.11.5
PyYAML==3.12
urllib3==1.22
